// src/data/enums/color.ts
export enum Color {
    color_red = 'color_red',
    color_pink = 'color_pink',
    color_orange = 'color_orange',
    color_yellow = 'color_yellow',
    color_mint = 'color_mint',
    color_green = 'color_green',
    color_cyan = 'color_cyan',
    color_blue = 'color_blue',
    color_dark_blue = 'color_dark_blue',
    color_indigo = 'color_indigo',
    color_purple = 'color_purple',
    color_brown = 'color_brown',
    color_white = 'color_white',
}